+++ src/lib/forms-schema/elements/CheckboxElement.ts
@@ -10,6 +10,7 @@ import {
   lookupSchemas,
   buttons,
   optionsSchemas,
+  canToggleAll,
 } from '../property-schemas'
 import { DYNAMIC_OPTION_TYPE } from '../common'
 export default Joi.object({
@@ -28,4 +29,5 @@ export default Joi.object({
   }),
   buttons,
   ...optionsSchemas,
+  canToggleAll,
 })