+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `canToggleAll` property to checkbox and multi select elements
+
 ## [0.9.13] - 2021-06-02
 
 ### Updated